+++ lib/elixir/test/elixir/range_test.exs
@@ -3,6 +3,11 @@ Code.require_file "test_helper.exs", __DIR__
 defmodule RangeTest do
   use ExUnit.Case, async: true
 
+  test :precedence do
+    assert Enum.to_list(1..3+2) == [1,2,3,4,5]
+    assert 1..3 |> Enum.to_list == [1,2,3]
+  end
+
   test :first do
     assert Range.new(first: 1, last: 3).first == 1
   end